Bundled tracker services move to core — core declares the appliance's own integrations
