+++ backend/tests/ship/test_ticketing.py
@@ -2,7 +2,6 @@ import json
 
 import httpx
 import pytest
-from druks.contrib.ship import services
 from druks.contrib.ship.extension import Ship, check_tracker_identity
 from druks.contrib.ship.ticketing.enums import TicketStatus
 from druks.contrib.ship.ticketing.exceptions import (
@@ -12,6 +11,7 @@ from druks.contrib.ship.ticketing.exceptions import (
 )
 from druks.contrib.ship.ticketing.jira import Jira, JiraClient
 from druks.contrib.ship.ticketing.linear import Linear, LinearClient
+from druks.core import services
 from druks.services import ServiceConnectError
 from druks.services.models import ServiceIdentity
